requirements.txt
watchdog
pystray
pillow
winshell
pywin32
python-docx
pypdf
